Create a post helper and implements it
